Extract menus to separate classes
